+++ backend/src/main/java/com/yigongil/backend/utils/querycounter/PreparedStatementProxyHandler.java
@@ -1,4 +1,4 @@
-package com.yigongil.backend.ui.querycounter;
+package com.yigongil.backend.utils.querycounter;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
