+++ backend/src/test/java/com/yigongil/backend/ui/StudyControllerTest.java
@@ -22,7 +22,7 @@ import com.yigongil.backend.fixture.MemberFixture;
 import com.yigongil.backend.request.StudyCreateRequest;
 import com.yigongil.backend.request.TodoCreateRequest;
 import com.yigongil.backend.request.TodoUpdateRequest;
-import com.yigongil.backend.ui.querycounter.ApiQueryCounter;
+import com.yigongil.backend.utils.querycounter.ApiQueryCounter;
 import java.util.Optional;
 import org.apache.http.HttpHeaders;
 import org.junit.jupiter.api.BeforeEach;
